feat: 가독성 좋게 설명 추가
